Changed camera yaw selected
Previously camera yaws for the straining and sliding frames were provided by the user. But, it's probably not going to be possible to know that ahead of time, since your freefall motion will influence the camera focus. So, now it finds all reasonable camera yaws from an input Lakitu position and checks them all.